fix ShopperInsightApiResultUnitTest
